# Web/NC.Web/NC.Web.Client/openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "NoobzCord API", "version": "1.0.0" },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "description": "Bad Request" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "operationId": "me",
        "security": [{ "bearer": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["contact", "password"],
        "properties": {
          "contact": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["name", "contact", "password"],
        "properties": {
          "name": { "type": "string", "minLength": 3 },
          "contact": { "type": "string" },
          "password": { "type": "string", "minLength": 6 }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "userId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "contact": { "type": "string" }
        }
      }
    }
  }
}