# Web/NC.Web/NC.Web.Client/src/api/generated/models/authResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NoobzCord API
 * OpenAPI spec version: 1.0.0
 */

export interface AuthResponse {
  token?: string;
  userId?: string;
  name?: string;
  contact?: string;
}