# Web/NC.Web/NC.Web.Client/src/api/generated/models/registerRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NoobzCord API
 * OpenAPI spec version: 1.0.0
 */

export interface RegisterRequest {
  /** @minLength 3 */
  name: string;
  contact: string;
  /** @minLength 6 */
  password: string;
}